- convert all models to json_ser - solve connection bug
