+++ lib/features/about_company/data/services/about_company_services.dart
@@ -2,7 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:dio/dio.dart';
 import 'package:galaxyxgate/core/errors/server_failure.dart';
 import 'package:galaxyxgate/core/networking/dio_helper.dart';
-import 'package:galaxyxgate/features/about_company/data/models/company_info_model.dart';
+import 'package:galaxyxgate/features/about_company/data/models/company_info/company_info.dart';
 
 class AboutCompanyServices {
   AboutCompanyServices({
