+++ lib/features/bottom_nav_bar/view/widgets/blured_bottom_nav_bar/blured_bottom_nav_bar.dart
@@ -49,7 +49,7 @@ class _BluredBottomNavBarState extends State<BluredBottomNavBar> {
   Widget build(BuildContext context) {
     return PopScope(
       canPop: false,
-      onPopInvoked: (_) async => false,
+      onPopInvoked: (_) async => bottom.value = baseBottom,
       child: Scaffold(
         body: Stack(
           children: [
