+++ lib/features/rockets/screens/rockets_screen.dart
@@ -5,7 +5,7 @@ import 'package:galaxyxgate/core/widgets/defult_app_header.dart';
 import 'package:galaxyxgate/core/widgets/gradient_gray_background.dart';
 import 'package:galaxyxgate/features/rockets/screens/widget/animated_rockets_listview.dart';
 import 'package:galaxyxgate/features/rockets/logic/rockets_cubit.dart';
-import 'package:galaxyxgate/features/rockets/data/models/rockets_model.dart';
+import 'package:galaxyxgate/features/rockets/data/models/rockets/rockets.dart';
 import 'package:galaxyxgate/features/rockets/screens/widget/place_holders/rockets_place_holder.dart';
 
 class RocketsScreen extends StatelessWidget {
@@ -46,7 +46,7 @@ class RocketsScreen extends StatelessWidget {
                             child: RocketsPlaceHolder());
                       }
                       if (state is RocketsLoaded) {
-                        List<RocketsModel>? rockets = state.rockets;
+                        List<Rockets>? rockets = state.rockets;
                         return AnimatedRocketsListView(
                           rockets: rockets!,
                         );
