+++ lib/features/ships/data/services/ships_services.dart
@@ -1,11 +1,11 @@
 import 'package:dio/dio.dart';
 import 'package:galaxyxgate/core/errors/server_failure.dart';
 import 'package:galaxyxgate/core/networking/dio_helper.dart';
-import 'package:galaxyxgate/features/ships/data/models/ships_model.dart';
+import 'package:galaxyxgate/features/ships/data/models/ships.dart';
 
 abstract interface class ShipsService {
-  Future<List<ShipsModel>> getAllShips();
-  Future<List<ShipsModel>> getFilteredShips();
+  Future<List<Ships>> getAllShips();
+  Future<List<Ships>> getFilteredShips();
 }
 
 class ShipsServiceImp implements ShipsService {
@@ -13,10 +13,10 @@ class ShipsServiceImp implements ShipsService {
   ShipsServiceImp({required this.dioHelper});
 
   @override
-  Future<List<ShipsModel>> getAllShips() async {
+  Future<List<Ships>> getAllShips() async {
     try {
       List<dynamic> response = await dioHelper.getRequest(endPoint: 'ships');
-      return response.map((ship) => ShipsModel.fromJson(ship)).toList();
+      return response.map((ship) => Ships.fromJson(ship)).toList();
     } catch (e) {
       if (e is DioException) {
         throw ServerFailure.fromDioException(dioException: e);
@@ -26,12 +26,12 @@ class ShipsServiceImp implements ShipsService {
   }
 
   @override
-  Future<List<ShipsModel>> getFilteredShips() async {
+  Future<List<Ships>> getFilteredShips() async {
     // First, fetch all ships using the existing method
-    final List<ShipsModel> ships = await getAllShips();
+    final List<Ships> ships = await getAllShips();
 
     // Filter ships to include only those with non-null images
-    final List<ShipsModel> filteredShips =
+    final List<Ships> filteredShips =
         ships.where((ship) => ship.image != null).toList();
 
     // Return the filtered list
