+++ lib/features/ships/screens/ships_bloc_provider.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:galaxyxgate/core/di/dependency_injection.dart';
 import 'package:galaxyxgate/features/favourits/logic/cubit/favourite_cubit.dart';
 import 'package:galaxyxgate/features/ships/logic/cubit/ships_cubit.dart';
-import 'package:galaxyxgate/features/ships/presentation/ships_screen.dart';
+import 'package:galaxyxgate/features/ships/screens/ships_screen.dart';
 
 class ShipsBlocProvider extends StatelessWidget {
   const ShipsBlocProvider({super.key});
