+++ lib/features/ships/screens/ships_screen.dart
@@ -5,8 +5,8 @@ import 'package:galaxyxgate/core/helpers/app_localization/app_localization.dart'
 import 'package:galaxyxgate/core/widgets/defult_app_header.dart';
 import 'package:galaxyxgate/core/widgets/gradient_gray_background.dart';
 import 'package:galaxyxgate/features/ships/logic/cubit/ships_cubit.dart';
-import 'package:galaxyxgate/features/ships/presentation/widget/animated_ships_list.dart';
-import 'package:galaxyxgate/features/ships/presentation/widget/place_holders/ships_place_holder.dart';
+import 'package:galaxyxgate/features/ships/screens/widget/animated_ships_list.dart';
+import 'package:galaxyxgate/features/ships/screens/widget/place_holders/ships_place_holder.dart';
 
 class ShipsScreen extends StatefulWidget {
   const ShipsScreen({super.key});
