+++ lib/features/ships/screens/widget/place_holders/ships_place_holder.dart
